LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/mix-blend-mode-only-on-transition-expected.txt


FAIL Blend modes are set up on paired transitions assert_equals: expected "isolate" but got "auto"
